requirements.txt
pdfplumber
streamlit
google-generativeai
python-dotenv
dotenv
google-generativeai
pdf2image